orders/requirements.txt
Flask==2.1.2
redis==4.3.4
requests==2.28.1
mysql-connector-python==8.0.30
Werkzeug<2.2